pyproject.toml
[tool.hatch.build.targets.wheel]
packages = ["condenses_validating"]

[tool.uv.sources]
subnet-node-managing = { git = "https://github.com/condenses/subnet-node-managing.git" }
text-compress-scoring = { git = "https://github.com/condenses/text-compress-scoring.git" }
restful-bittensor = { git = "https://github.com/condenses/restful-bittensor.git" }
subnet-synthesizing = { git = "https://github.com/condenses/subnet-synthesizing" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "subnet-validating"
version = "0.1.0"
description = "Subnet Validator"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "bittensor==9.0.0",
    "fastapi",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.7.1",
    "pymongo>=4.11.1",
    "redis>=5.2.1",
    "restful-bittensor",
    "subnet-node-managing",
    "subnet-synthesizing",
    "text-compress-scoring",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/restful-bittensor"
Repository = "https://github.com/yourusername/restful-bittensor.git"


[project.scripts]
condenses-validating = "condenses_validating.main:start_loop"
